Fix lookup terminfo entries outside nix
We want this standalone binary to run outside of the nix environment
that produced it, where the nix store is not available. The default
directory to lookup for entries for terminfo is set to be in the nix
store. We take advantage of the TERMINFO_DIRS environment variable to
lookup entries in more places where terminfo definitions usually exist.

More information:

* man terminfo: Fetching Compiled Descriptions

* cat /etc/terminfo/README:

```
This directory is for system-local terminfo descriptions. By default,
ncurses will search ${HOME}/.terminfo first, then /etc/terminfo (this
directory), then /lib/terminfo, and last not least /usr/share/terminfo.
```