Set terminal background color to match solarized
Set the terminal background color to match the color used by the color
scheme for the default attribute.

We do this because in the edge of the terminal you get some of the
default color and it does not look nice, especially if there is high
contrast.

Add commit to vty: "Add ability to set terminal background color"