Add Libgit2 FFI for signatures and configuration
Configuration is not going to be needed right now but added it anyways.
There is code to find out the correct signature out of the configuration
already in the signature.h file.

Add missing finalizers for Diff(Find)Options.

Change Signature to newtype from structures:

It is more complicated and less efficient to keep it as a Haskell
structure. More complicated because we need to provide the poke function
as well and less efficient because even if we create a signature just to
pass it to another function we need to both peek and poke for no reason.
Hybrid solutions are even more complicated.

We provide two way to read signatures from C depending if the ownership
of the object is within Libgit2 or with the client code.