Use tinted-theming/base16-schemes
All themes are collected in one place and so it is more likely that this
will build in the future. The original base16 theme list seems to be a
bit out of date and several themes are not available on GitHub anymore.
Also this way I don't need to maintain an extra repo collecting all the
themes and packaging them with nix.

When I move to a more recent nixpkgs, I should use another builder for
the themes, one of the recommended from tinted-theming:

https://github.com/tinted-theming/home#builders

There is also more modern providers for base16 for nix, e.g.:

https://github.com/SenchoPens/base16.nix/tree/main

but I am not sure if they are going to be useful to this project.